+++ R/plot.R
@@ -319,3 +319,36 @@ plot_crop = function(x, quiet = !opts_knit$get('progress')) {
 showtext = function(show) {
   if (isTRUE(show)) getFromNamespace('showtext.begin', 'showtext')()
 }
+
+#' Embed external images in \pkg{knitr} documents
+#'
+#' When plots are not generated from R code, there is no way for \pkg{knitr} to
+#' capture plots automatically. In this case, you may generate the images
+#' manually and pass their file paths to this function to include them in the
+#' output. The major advantage of using this function is that it is portable in
+#' the sense that it works for all document formats that \pkg{knitr} supports,
+#' so you do not need to think if you have to use, for example, LaTeX or
+#' Markdown syntax, to embed an external image. Chunk options related to
+#' graphics output that work for normal R plots also work for these images, such
+#' as \code{out.width} and \code{out.height}.
+#' @param path a character vector of image paths
+#' @param auto_pdf whether to use PDF images automatically when the output
+#'   format is LaTeX, e.g. \file{foo/bar.png} will be replaced by
+#'   \file{foo/bar.pdf} if the latter exists; this can be useful since normally
+#'   PDF images are of higher qualities than raster images like PNG when the
+#'   output is LaTeX/PDF
+#' @note This function is supposed to be used in R code chunks or inline R code
+#'   expressions. You are recommended to use forward slashes (\verb{/}) as path
+#'   separators instead of backslashes in the image paths.
+#' @return The same as the input character vector \code{path} but it is marked
+#'   with special internal S3 classes so that \pkg{knitr} will convert the file
+#'   paths to proper output code according to the output format.
+#' @export
+include_graphics = function(path, auto_pdf = TRUE) {
+  if (auto_pdf && is_latex_output()) {
+    path2 = sub_ext(path, 'pdf')
+    i = file.exists(path2)
+    path[i] = path2[i]
+  }
+  structure(path, class = c('knit_image_paths', 'knit_asis'))
+}